Created new GenericStyles.ts file which will hold all the regularly reused css throughout MMA. First refactor has been moving sectionSpacing into GenericStyles.ts. Same css was being created in three separate style files so they have been removed and new location has been imported into the relevant places. This should have no breaking or UI changes.
