+++ client/components/mma/cancel/cancellationSaves/SaveOptions.tsx
@@ -16,6 +16,7 @@ import { Navigate, useLocation, useNavigate } from 'react-router';
 import type { PaidSubscriptionPlan } from '../../../../../shared/productResponse';
 import { getMainPlan } from '../../../../../shared/productResponse';
 import { calculateMonthlyOrAnnualFromBillingPeriod } from '../../../../../shared/productTypes';
+import { sectionSpacing } from '../../../../styles/GenericStyles';
 import {
 	getNewMembershipPrice,
 	getOldMembershipPrice,
@@ -38,7 +39,6 @@ import {
 	headingCss,
 	productSubtitleCss,
 	productTitleCss,
-	sectionSpacing,
 } from './SaveStyles';
 
 const NewPriceIcon = () => {
