+++ client/components/mma/switch/review/SwitchReview.tsx
@@ -13,6 +13,7 @@ import { Navigate, useLocation, useNavigate } from 'react-router';
 import { Link } from 'react-router-dom';
 import { dateString } from '../../../../../shared/dates';
 import type { ProductSwitchType } from '../../../../../shared/productSwitchTypes';
+import { sectionSpacing } from '../../../../styles/GenericStyles';
 import {
 	LoadingState,
 	useAsyncLoader,
@@ -40,7 +41,6 @@ import {
 	iconListCss,
 	listWithDividersCss,
 	productTitleCss,
-	sectionSpacing,
 	smallPrintCss,
 } from '../SwitchStyles';
